Conflicts:
	UIKit/UIKit.xcodeproj/project.pbxproj
	UIKit/UIKit.xcodeproj/xcshareddata/xcschemes/UIKit.xcscheme

+++ UIKit/Classes/UIActivityIndicatorView.m
@@ -77,7 +77,7 @@ static UIImage *UIActivityIndicatorViewFrameImage(UIActivityIndicatorViewStyle s
         
         // position and draw the tooth
         CGContextRotateCTM(c, 1 / numberOfTeeth * TWOPI);
-        [[UIBezierPath bezierPathWithRoundedRect:CGRectMake(-toothWidth/2.f,-radius,toothWidth,ceil(radius*.54f)) cornerRadius:toothWidth/2.f] fill];
+        [[UIBezierPath bezierPathWithRoundedRect:CGRectMake(-toothWidth/2.f,-radius,toothWidth,ceilf(radius*.54f)) cornerRadius:toothWidth/2.f] fill];
     }
     
     // hooray!
